src/graph_rag/rag_code/requirements.txt
transformers>=4.40.0
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
sentence-transformers>=3.0.0

langchain-core==0.3.71
langchain-community==0.3.27
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.8
unstructured-client==0.41.0
neo4j>=5.0.0

pymilvus==2.5.11
rank-bm25>=0.2.2

lazy_loader==0.4
huggingface-hub>=0.33.4
tokenizers>=0.19.0
accelerate>=0.20.0

openai>=1.86.0,<2.0.0

tiktoken>=0.4.0

python-dotenv>=1.0.0

numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
tqdm>=4.64.0
pydantic>=2.0.0
requests>=2.28.0
faiss 